requirements.txt
pyyaml>=6.0.1,<7
pydantic>=2.7,<3
typer>=0.12,<0.13
rich>=13.7,<14
robocode-tank-royale==0.34.1
websockets>=11,<16
